Generate a client for the MachineConfig types
In order to manipulate the MachineConfig types we defined in the
previous commit, we need to generate a client. To do that, we're going
to use the "k8s.io/code-generator/cmd/client-gen" project.

The main integration point is the "gen-mcfg-client" Makefile target
that takes the API definition under
github.com/openshift/compliance-operator/pkg/apis and generates the
client into github.com/openshift/compliance-operator/pkg/generated.

The custom-boilerplate.go.txt is just a dummy file the generator needs.